Renamed Components in proper way
