+++ app/admin/forgot-password/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ForgotpasswordFlow from "@/components/admin/forgotPasswordFlow/forgotpasswordflow";
+import ForgotpasswordFlow from "@/components/admin/forgotPasswordFlow/ForgotPasswordFlow";
 
 const Forgotpasswordpage = () => {
   return (
